Combat feature

+++ src/Equipment.java
@@ -4,15 +4,15 @@ import java.io.Serializable;
 public class Equipment extends Item implements Serializable {
 
     private int equipmentSlot;
-    private int damageReduction;
+    private float damageReduction;
 
-    public Equipment(String name, int itemID, String description, int equipmentSlot, int damageReduction) {
+    public Equipment(String name, int itemID, String description, int equipmentSlot, float damageReduction) {
         super(name, itemID, description);
         this.equipmentSlot = equipmentSlot;
         this.damageReduction = damageReduction;
     }
 
-    public int getDamageReduction() {
+    public float getDamageReduction() {
         return damageReduction;
     }
     public int getEquipmentSlot(){
